### 2025-10-13 #9226 @editorialbot
[PRE REVIEW]: ImageMLResearch: A Python Toolkit for Reproducible Image-Based ML Experiments
**Submitting author:** @SiulRek (<a href="http://orcid.org/0009-0009-8692-9227">Luis Kraker</a>)
**Repository:** https://github.com/SiulRek/ImageMLResearch
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.0
