### 2025-10-13 223d714
Add missing DOIs based on JOSS editorial bot feedback
