+++ paper/paper.bib
@@ -30,7 +30,8 @@
   number={3},
   pages={56--68},
   year={2018},
-  publisher={Association for the Advancement of Artificial Intelligence}
+  publisher={Association for the Advancement of Artificial Intelligence},
+  doi={10.1609/aimag.v39i3.2816}
 }
 
 @article{pineau2021improving,
@@ -51,7 +52,8 @@
   number={6377},
   pages={725--726},
   year={2018},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.359.6377.725}
 }
 
 @inproceedings{paszke2019pytorch,
