### 2025-10-14 3ae6fd3
Replace jpg with png, remove "ENDLESS" from affiliation
