### 2025-10-14 dbd2c33
Enhance image title clarity and consistency
