### 2025-11-04 #9226 @crvernon
👋 @SiulRek - Thank your for your submission to JOSS. We are currently managing a large backlog of submissions but will get you a topic editor assigned as soon as possible.

For now, we will need to waitlist to designate that it is ready when a topic editor becomes available. Thanks for your patience!
