### 2026-02-03 23d0b1a
Update paper.md to meet JOSS submission requirements
