### 2026-02-03 389173b
Update paper.md to meet JOSS submission requirements
