+++ paper/paper.md
@@ -24,24 +24,29 @@ bibliography: paper.bib
 
 # ImageMLResearch: A Python Toolkit for Reproducible Image-Based ML Experiments
 
+## Summary
+
 `ImageMLResearch` is an open-source Python toolkit that streamlines and standardizes image-based machine learning (ML) research. While ML has achieved remarkable success in computer vision, the complexity of research workflows remains a barrier to reproducibility and accessibility. Many projects rely on loosely connected scripts or notebooks, leading to fragmented experiment management and limited reproducibility.
 
 `ImageMLResearch` addresses this gap by providing a modular Python package with a clear API, without requiring intrusive dashboards or command-line interfaces. Built on widely adopted libraries such as TensorFlow [@abadi2016tensorflow], Keras [@chollet2015keras], and Optuna [@akiba2019optuna], it offers a lightweight, research-oriented approach to reproducible image-based ML experimentation. The toolkit is designed to support education, exploratory research, and the development of more robust experiment management practices.
 
 ## Statement of Need
-Image-based machine learning workflows are often constructed from ad hoc scripts or notebooks, making it difficult to maintain a clear structure between data handling, preprocessing, training, and evaluation. This fragmentation contributes to poor reproducibility and hinders systematic experimentation [@gundersen2018reproducible; @pineau2021improving; @hutson2018reproducibility]. 
+Image-based machine learning workflows are often constructed from ad hoc scripts or notebooks, making it difficult to maintain a clear structure between data handling, preprocessing, training, and evaluation. This fragmentation contributes to poor reproducibility and hinders systematic experimentation [@gundersen2018reproducible; @pineau2021improving; @hutson2018reproducibility].
 
-General-purpose frameworks such as TensorFlow [@abadi2016tensorflow] and PyTorch [@paszke2019pytorch] provide the computational foundations but do not prescribe experiment organization. Tools such as MLflow [@zaharia2018mlflow] and Weights & Biases [@biewald2020wandb] extend functionality with experiment tracking and visualization, yet they often require additional infrastructure that can be burdensome in lightweight academic projects.
+While modern machine learning libraries provide powerful computational building blocks, they do not enforce a coherent structure for managing experiments. As a result, researchers must manually coordinate configurations, results, and documentation, which increases cognitive overhead and the likelihood of irreproducible outcomes.
 
-`ImageMLResearch` addresses this gap by structuring the experiment lifecycle for image classification into modular components. It supports the consistent definition, execution, and documentation of experiments without reliance on external services. This makes the toolkit especially suited for exploratory research, and smaller projects where transparent and reproducible experiment management is essential.
+`ImageMLResearch` was developed to address these challenges by providing a lightweight, structured framework for defining, executing, and documenting image-based machine learning experiments in a reproducible manner.
 
 ## State of the Field
 
-Reproducibility and systematic experimentation are longstanding challenges in image-based machine learning research. Existing tools such as experiment tracking platforms (e.g., MLflow, Weights & Biases) and workflow managers provide infrastructure for logging metrics and managing runs, but they are often agnostic to domain-specific requirements in image analysis. As a result, researchers frequently rely on custom scripts and ad hoc conventions to manage datasets, preprocessing pipelines, model variants, and evaluation protocols, which complicates reproducibility and comparison across experiments.
+A variety of tools exist to support machine learning experimentation and reproducibility. Core frameworks such as TensorFlow [@abadi2016tensorflow] and PyTorch [@paszke2019pytorch] provide flexible abstractions for model development and training but leave experiment organization and result management largely to the user.
 
[…]
