### 2026-02-03 #9226 @schappag
We have updated paper.md to address the JOSS submission requirements, including the required paper sections and AI usage disclosure. Please let us know if any further changes are needed.
