### 2026-02-11 #9226 @schappag
Hi @openjournals/joss-eics,

Just a brief follow-up. The paper has been updated (including AI disclosure), the PDF proof was generated successfully, and reviewer suggestions have been provided.

Could you please let us know whether there is any further information needed from our side, or if an editor might be available to take this submission?

Thank you very much.
