### 2026-02-24 SiulRek/ImageMLResearch#1 @nkrusch
Contribution guidelines
Ref: https://github.com/openjournals/joss-reviews/issues/9226

add community guidelines, with clear guidelines for third parties wishing to 1. contribute to the software 2. report issues or problems with the software 3. seek support.
