### 2026-02-24 #9226 @nkrusch
👋  Hello @masadcv, @michaelberks, @Amorfati123, @sibocw and @pr4deepr,

would you be interested in reviewing this submission for JOSS?
