### 2026-02-25 #9226 @Amorfati123
Hi, thanks for reaching out. Yes, I'd be happy to review this paper. I'm interested and available to take this on. Please let me know the timeline and any next steps.
