### 2026-02-25 #9226 @nkrusch
Great, thanks @Amorfati123!

We are currently in the pre-review stage, where I (the editor) am looking for reviewers. After I have confirmed 2 (ideally 3) reviewers, we will move to the review phase.

Once the review starts, the goal is that reviewers complete a first pass of the submission in 2-4 weeks. The JOSS review follows a [checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist) and is iterative. When a reviewer discover criteria that is not satisfied, the author will make modifications to resolve the issues. The total time to completing reviews thus depends on this iteration. You can find more information about the JOSS reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

For now, you simply wait. You will see another notification when the review starts.
