### 2026-02-25 #9226 @sibocw
> 👋  Hello @masadcv, @michaelberks, @Amorfati123, @sibocw and @pr4deepr,
>
> would you be interested in reviewing this submission for JOSS?

Yes I'm happy to!
