### 2026-02-28 2d9118b
ci: Add GitHub Actions workflow for testing
Run test suite on push/PR to main and joss-paper branches.
Tests Python 3.10 and 3.12 on Ubuntu with pip caching.
