+++ .github/workflows/ci.yml
@@ -1,0 +1,34 @@
+name: CI
+
+on:
+  push:
+    branches: [main, joss-paper]
+  pull_request:
+    branches: [main, joss-paper]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.12"]
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: pip
+
+      - name: Install package
+        run: pip install .
+
+      - name: Run tests
+        run: |
+          python -c "
+          from imlresearch.api.tests import run_all_tests
+          result = run_all_tests()
+          exit(0 if result.wasSuccessful() else 1)
+          "
