### 2026-02-28 28aad14
docs: Add CONTRIBUTING.md with development guidelines
