### 2026-03-04 #10130 @Amorfati123
My comments:
1. For the Installation: All 8 dependencies are pinned to exact versions (e.g., tensorflow==2.17.0, pandas==2.2.2, matplotlib==3.8.0). This is poor practice for a library and virtually guarantees conflicts in any non-empty environment. The README itself acknowledges this and recommends a clean virtual environment. openai==1.34.0 is a hard dependency for a "proof-of-concept" AI analysis feature. This should be an optional dependency (extras_require), not forced on all users. I would suggest using version ranges (e.g., tensorflow>=2.15,<3.0) for library compatibility. Make openai an optional extra.

2. No test coverage reporting, consider adding coverage reporting.

3. I couldn't find CODE_OF_CONDUCT.md on any branch.

4. In Software Design section, the paper does not deeply discuss trade-offs (e.g., why TensorFlow-only and not also PyTorch? Why file-based outputs vs. a database? Why JSON for configs vs. YAML or Python?). These are the "meaningful design thinking" signals, so consider adding 2-3 sentences on key architectural trade-offs.

5. The References section in paper.bib does not include DOIs for all entries (e.g., Paszke et al. 2019 PyTorch entry lacks a DOI).
