### 2026-03-04 SiulRek/ImageMLResearch#7 @nkrusch
Pinned dependencies
Ref: https://github.com/openjournals/joss-reviews/issues/10130

For the Installation: All 8 dependencies are pinned to exact versions (e.g., tensorflow==2.17.0, pandas==2.2.2, matplotlib==3.8.0). This is poor practice for a library and virtually guarantees conflicts in any non-empty environment. The README itself acknowledges this and recommends a clean virtual environment. openai==1.34.0 is a hard dependency for a "proof-of-concept" AI analysis feature. This should be an optional dependency (extras_require), not forced on all users. I would suggest using version ranges (e.g., tensorflow>=2.15,<3.0) for library compatibility. Make openai an optional extra.
