### 2026-03-04 SiulRek/ImageMLResearch#8 @nkrusch
Test Coverage
Ref: https://github.com/openjournals/joss-reviews/issues/10130

No test coverage reporting, consider adding coverage reporting.
