### 2026-03-04 SiulRek/ImageMLResearch#9 @nkrusch
Add code of conduct
Ref: https://github.com/openjournals/joss-reviews/issues/10130

I couldn't find CODE_OF_CONDUCT.md on any branch.
