### 2026-03-04 #10130 @nkrusch
@Amorfati123, I have opened comments 1--3 as issues at the repository, since they are technical changes. @SiulRek, also note comments 4--5 about the paper. Paper comments usually get handled in this review thread (but I can open issues about those also, if it helps to track progress).
