### 2026-03-08 ab60838
Relax dependency constraints, make OpenAI optional
