+++ requirements.txt
@@ -1,8 +1,5 @@
-tensorflow==2.17.0
-pandas==2.2.2
-matplotlib==3.8.0
-openai==1.34.0
-optuna==3.6.1
-seaborn==0.13.2
-scikit-learn==1.4.1.post1
-opencv-python==4.8.1.78
+tensorflow>=2.13,<2.18
+optuna>=3.3,<4.8
+opencv-python>=4.8,<4.14
+scikit-learn>=1.4,<1.8
+seaborn>=0.12,<0.14
