+++ pyproject.toml
@@ -12,6 +12,7 @@ authors = [
     { name = "SiulRek", email = "luiskraker2000@gmail.com" }
 ]
 dynamic = ["version"]
+
 classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
@@ -27,17 +28,17 @@ classifiers = [
 ]
 urls = { "Repository" = "https://github.com/SiulRek/ImageMLResearch", "Documentation" = "https://imagemlresearch.readthedocs.io/en/api-development/" }
 
-
-
 dependencies = [
-    "tensorflow==2.17.0",
-    "pandas==2.2.2",
-    "matplotlib==3.8.0",
-    "openai==1.34.0",
-    "optuna==3.6.1",
-    "seaborn==0.13.2",
-    "scikit-learn==1.4.1.post1",
-    "opencv-python==4.8.1.78"
+    "tensorflow>=2.13,<2.18",
+    "optuna>=3.3,<4.8",
+    "opencv-python>=4.8,<4.14",
+    "scikit-learn>=1.4,<1.8",
+    "seaborn>=0.12,<0.14",
+]
+
+[project.optional-dependencies]
+ai-report = [
+    "openai==1.34.0"
 ]
 
 [tool.setuptools.dynamic]
@@ -49,7 +50,6 @@ where = ["."]
 [tool.setuptools.package-data]
 # Include required non-Python files (e.g., test data)
 "*" = ["*.png", "*.jpg", "*.tfrecord", "*.json"]
-
 
 # Apply ruff checks to imlresearch/src
 [tool.ruff]
@@ -73,7 +73,6 @@ ignore = [
 
 [tool.ruff.lint.isort]
 force-single-line = true
-
 
 [tool.pylint.'MAIN']
 load-plugins = ["pylint_per_file_ignores"]
@@ -102,7 +101,5 @@ per-file-ignores = [
     "imlresearch/src/preprocessing/steps/*:W0613"
 ]
 
-
 [tool.black]
 line-length = 79
-
