### 2026-03-08 4723dca
Add NumPy<2 constraint to avoid TensorFlow import errors
