+++ requirements.txt
@@ -1,3 +1,4 @@
+numpy>=1.23.5,<2
 tensorflow>=2.13,<2.18
 optuna>=3.3,<4.8
 opencv-python>=4.8,<4.14
