### 2026-03-08 SiulRek/ImageMLResearch#7 @SiulRek
Relaxed dependency constraints and made `openai` optional in [commit `4723dca`](https://github.com/SiulRek/ImageMLResearch/commit/4723dca4715bbfe75ee4961284bd57a9e8828d7d) and [commit `ab60838`](https://github.com/SiulRek/ImageMLResearch/commit/ab60838b0ea10c6571be2b730f1b75560ccffb7d).
