### 2026-03-08 c85e8fb
Add test coverage execution in build process
