### 2026-03-08 bf8ae89
Add CI checks for ruff, tests, and coverage
