+++ .github/workflows/ci.yml
@@ -10,6 +10,7 @@ jobs:
   test:
     runs-on: ubuntu-latest
     strategy:
+      fail-fast: false
       matrix:
         python-version: ["3.10", "3.12"]
 
@@ -22,13 +23,21 @@ jobs:
           python-version: ${{ matrix.python-version }}
           cache: pip
 
-      - name: Install package
-        run: pip install .
-
-      - name: Run tests
+      - name: Install package and CI tools
         run: |
-          python -c "
-          from imlresearch.api.tests import run_all_tests
-          result = run_all_tests()
-          exit(0 if result.wasSuccessful() else 1)
-          "
+          python -m pip install --upgrade pip
+          pip install .
+          pip install ruff coverage
+
+      - name: Run checks
+        run: |
+          python <<'PY'
+          from pip_build import build_package
+
+          try:
+              build_package(check_only=True)
+          except Exception:
+              raise SystemExit(1)
+          else:
+              raise SystemExit(0)
+          PY
