### 2026-03-08 cefb953
Update command in run_test_coverage function
