### 2026-03-08 SiulRek/ImageMLResearch#8 @SiulRek
Test coverage reporting has been added to the GitHub CI workflow using the coverage package. The CI checks that total statement and branch coverage is at least 80%.
