### 2026-03-19 7f757d1
Update bibliography reference metadata
