+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={TensorFlow: Large-scale machine learning on heterogeneous distributed systems},
   author={Abadi, Mart{\'i}n and Agarwal, Ashish and Barham, Paul and Brevdo, Eugene and Chen, Zhifeng and Citro, Craig and Corrado, Greg S. and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and others},
   journal={arXiv preprint arXiv:1603.04467},
-  year={2016}
+  year={2016},
+  doi={10.48550/arXiv.1603.04467}
 }
 
 @misc{chollet2015keras,
@@ -19,7 +20,8 @@
   booktitle={Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
   pages={2623--2631},
   year={2019},
-  organization={ACM}
+  organization={ACM},
+  doi={10.1145/3292500.3330701}
 }
 
 @article{gundersen2018reproducible,
@@ -37,11 +39,9 @@
 @article{pineau2021improving,
   title={Improving reproducibility in machine learning research (a report from the NeurIPS 2019 reproducibility program)},
   author={Pineau, Joelle and Vincent-Lamarre, Philippe and Sinha, Koustuv and Larivi{\`e}re, Vincent and Beygelzimer, Alina and d'Alch{\'e}-Buc, Florence and Fox, Emily and Larochelle, Hugo},
-  journal={Journal of Machine Learning Research},
-  volume={22},
-  number={164},
-  pages={1--20},
-  year={2021}
+  journal={arXiv preprint arXiv:2003.12206},
+  year={2021},
+  doi={10.48550/arXiv.2003.12206}
 }
 
 @article{hutson2018reproducibility,
@@ -61,21 +61,25 @@
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
   booktitle={Advances in Neural Information Processing Systems},
   volume={32},
-  year={2019}
+  year={2019},
+  doi={10.5555/3454287.3455008}
 }
 
-@inproceedings{zaharia2018mlflow,
-  title={Accelerating the machine learning lifecycle with MLflow},
-  author={Zaharia, Matei and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, S. Andy and Konwinski, Andy and Murching, Sourabh and Nykodym, Tomas and Ogilvie, Michael and Parkhe, Manish and others},
-  booktitle={Proceedings of the 4th International Workshop on Data Management for End-to-End Machine Learning},
-  pages={39--44},
+@article{zaharia2018mlflow,
+  title={Accelerating the Machine Learning Lifecycle with MLflow},
+  author={Zaharia, Matei and Chen, Andrew and Davidson, Aaron and Ghodsi, Ali and Hong, Sue Ann and Konwinski, Andy and Murching, Siddharth and Nykodym, Tomas and Ogilvie, Paul and Parkhe, Mani and Xie, Fen and Zumar, Corey},
+  journal={IEEE Data Engineering Bulletin},
+  volume={41},
+  number={4},
+  pages={39--45},
   year={2018},
-  organization={ACM}
+  url={https://people.eecs.berkeley.edu/~matei/papers/2018/ieee_mlflow.pdf}
 }
 
 @misc{biewald2020wandb,
   title={Experiment tracking with Weights and Biases},
   author={Biewald, Lukas},
   year={2020},
-  howpublished={Software available from \url{https://wandb.com}}
[…]
