### 2026-03-20 8852a59
Add design thinking insights in paper.md
