### 2026-03-20 #10130 @SiulRek
@Amorfati123, Thank you for the helpful suggestions. We have clarified the key architectural choices in the software design section and added DOIs to all references where available. Some sources do not have a DOI. Note, comments 1-3 have been addressed in the repository issues.
