### 2026-03-21 SiulRek/ImageMLResearch#9 @nkrusch
at [joss-paper/CODE_OF_CONDUCT.md](https://github.com/SiulRek/ImageMLResearch/blob/joss-paper/CODE_OF_CONDUCT.md)
