### 2026-03-21 #10130 @nkrusch
@SiulRek please check this entry in the bib

> - 10.5555/3454287.3455008 is INVALID - 10.5555 is a known broken prefix, replace with https://dl.acm.org/doi/{doi} in > the {url} field
