### 2026-03-21 1558346
Replace broken 10.5555 DOI with ACM DL URL in paper.bib
