+++ paper/paper.bib
@@ -62,7 +62,7 @@
   booktitle={Advances in Neural Information Processing Systems},
   volume={32},
   year={2019},
-  doi={10.5555/3454287.3455008}
+  url={https://dl.acm.org/doi/10.5555/3454287.3455008}
 }
 
 @article{zaharia2018mlflow,
