### 2026-04-02 #10130 @sibocw
Sorry for the delay. I've updated my checklist (everything is checked now).

My comments:
1. It's a bit counterintuitive to me that the tests are under `imlresearch/src/testing/`. Is there a reason why it's not just in a `tests/` folder at the root level?
2. Any plan to set up CI/CD for automated testing whenever a commit is pushed to GitHub? This shouldn't be very hard to set up but can be very useful, especially if people who are less familiar with/confident about the codebase want to contribute without having to worry about breaking everything.

My nitpicks:
1. References formatting: Software citations have different styles (i.e. Keras vs. W&B). To me this is perfectly acceptable but I think the editors should decide whether this is fine.
2. References heading: The bibliography seems to be under the Acknowledgment section. Is there a missing section heading, or is this just journal style?
3. `pyproject.toml`: `requires-python` is `>=3.10`, but Python 3.13/3.14 are missing from `classifiers`.
4. `pyproject.toml`: I would personally prefer to pin seaborn to something stricter, for example `>=0.12.0,<=0.13.2` (if you have verified that it works). Per [Semantic Versioning] rules, "Major version zero (0.y.z) is for initial development. Anything MAY change at any time. The public API SHOULD NOT be considered stable."
