### 2026-04-07 #10130 @nkrusch
@SiulRek please respond to the [comments](https://github.com/openjournals/joss-reviews/issues/10130#issuecomment-4178415916)

Regarding references formatting:
After the reviewers are done, I will do my round of editorial checks that includes checking the bibliography. As a rule of thumb, bibliographic entries of the same kind should be consistent in style.
