### 2026-04-07 395efaa
Clean and standardize bibliography entries in paper.bib
