+++ paper/paper.bib
@@ -7,11 +7,11 @@
 }
 
 @misc{chollet2015keras,
-  title={Keras},
-  author={Chollet, Fran{\c{c}}ois and others},
-  year={2015},
-  howpublished={GitHub repository},
-  url={https://github.com/keras-team/keras}
+  author = {Chollet, François},
+  title = {Keras},
+  year = {2015},
+  url = {https://github.com/keras-team/keras},
+  note = {Software}
 }
 
 @inproceedings{akiba2019optuna,
@@ -77,9 +77,9 @@
 }
 
 @misc{biewald2020wandb,
-  title={Experiment tracking with Weights and Biases},
-  author={Biewald, Lukas},
-  year={2020},
-  howpublished={Software available from \url{https://wandb.ai/site/experiment-tracking/}},
-  url={https://wandb.ai/site/experiment-tracking/}
+  author = {Biewald, Lukas},
+  title = {Weights \& Biases},
+  year = {2020},
+  url = {https://wandb.ai/site/experiment-tracking/},
+  note = {Software}
 }
