### 2026-04-07 4beb60d
Restrict seaborn dependency to 0.12.0-0.13.2 range
