### 2026-04-07 6dbc91c
Extracted run_pre_build_checks method, update CI
