### 2026-04-07 #10130 @SiulRek
@sibocw Many thanks for your helpful review.

Regarding the comments:
  1. Our reasoning is that all module-specific tests already live within their respective module folders. The src/testing/ directory is only used for shared test utilities, base classes, and the master test runner. Moving this to a root-level tests/ folder would mostly result in a directory containing just that runner and helpers, while the actual tests would still remain within the modules.
  2. We updated ci.yaml to make this clear. When changes are pushed to the repository, our GitHub Action defined in ci.yaml automatically runs code style checks, unit tests, and test coverage checks on every push, so contributors can immediately see if something is broken.

Regarding the minor issues:
  1. Thanks. We updated the References accordingly.
  2. As far as we understood this is journal style. But we are happy to change it if we are mistaken.
  3. Thanks. We haven’t tested compatibility with Python 3.13/3.14 yet. For now, we’ll restrict the range to >=3.10,<3.13 to reflect the versions we’ve actually verified, and update it once newer versions are tested.
  4. Thanks for the suggestion. We pinned it more strictly (>=0.12.0,<=0.13.2) based on the versions we’ve tested.
