### 2026-04-11 #10130 @sibocw
@SiulRek Thanks for the update! I understand the choice with test structuring now, and thanks for the CI update.

@nkrusch I'm fully satisfied now! Thanks.
