### 2026-04-21 #10130 @nkrusch
@SiulRek I have these actionable comments about the paper:

* Use top-level heading (`#` not `##`) for section headings
* Bibliography needs a title (add `# References` at the end of paper.md file)
* Illustrative Example/Figure 1: the image is pixelated; ideally use svg or pdf image format, or substitute the current image with a higher DPI one.
* Quality control: consider recognizing Pylint and Ruff repositories with a citation.
* Bibliography: TensorFlow-paper was published in the [proceedings of OSDI 16](https://dl.acm.org/doi/10.5555/3026877.3026899), but the paper citation refers to a preprint. Double-check which version you wanted to cite.
* Bibliography: Keras: add software version number.

Let me know when you are done, so I can proceed.
