### 2026-04-21 #10130 @nkrusch
@SiulRek, I also have a few side comments, independently of the JOSS review.

I would recommend that you enhance the ImageMLResearch workflow and [output logging](https://github.com/SiulRek/ImageMLResearch/tree/joss-paper/demo/mnist_digits/automatic/output) by capturing details of the executing environment (host OS, architecture, CPU/GPU, dependent software versions,...). This is important information for reproducing experiments.

How does ImageMLResearch handle software changes over time? One real challenge with reusing research software, or reproducing experiments, is handling the dependency drift over time. You can address this challenge in part by capturing the environment details carefully; but in general, it would be good to prepare for preservation and recovery of the experiment environment in advance.

I do not expect that you address these now; rather consider these over time as you develop the software.
