### 2026-04-22 326fd04
Replace pixelated Figure 1 with higher-quality version
