### 2026-04-22 8b1decd
Update Keras and TensorFlow bibliography entries
