+++ paper/paper.bib
@@ -1,18 +1,22 @@
-@article{abadi2016tensorflow,
-  title={TensorFlow: Large-scale machine learning on heterogeneous distributed systems},
-  author={Abadi, Mart{\'i}n and Agarwal, Ashish and Barham, Paul and Brevdo, Eugene and Chen, Zhifeng and Citro, Craig and Corrado, Greg S. and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and others},
-  journal={arXiv preprint arXiv:1603.04467},
-  year={2016},
-  doi={10.48550/arXiv.1603.04467}
+@inproceedings{abadi2016tensorflow,
+  title     = {TensorFlow: A System for Large-Scale Machine Learning},
+  author    = {Abadi, Mart{\'i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and Kudlur, Manjunath and Levenberg, Josh and Monga, Rajat and Moore, Sherry and Murray, Derek G. and Steiner, Benoit and Tucker, Paul and Vasudevan, Vijay and Warden, Pete and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
+  booktitle = {12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)},
+  pages     = {265--283},
+  year      = {2016},
+  publisher = {USENIX Association},
+  url       = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi}
 }
 
 @misc{chollet2015keras,
-  author = {Chollet, François},
-  title = {Keras},
-  year = {2015},
-  url = {https://github.com/keras-team/keras},
-  note = {Software}
+  author  = {Chollet, Fran{\c{c}}ois and others},
+  title   = {Keras},
+  year    = {2015},
+  version = {3.12.1},
+  url     = {https://github.com/keras-team/keras},
+  note    = {Software}
 }
+
 
 @inproceedings{akiba2019optuna,
   title={Optuna: A next-generation hyperparameter optimization framework},
