### 2026-04-22 ede6e1e
Add ruff and pylint to bibliography
