### 2026-04-22 61212ad
Add ruff and pylint to bibliography
