+++ paper/paper.bib
@@ -1,89 +1,104 @@
 @inproceedings{abadi2016tensorflow,
-  title     = {TensorFlow: A System for Large-Scale Machine Learning},
-  author    = {Abadi, Mart{\'i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and Kudlur, Manjunath and Levenberg, Josh and Monga, Rajat and Moore, Sherry and Murray, Derek G. and Steiner, Benoit and Tucker, Paul and Vasudevan, Vijay and Warden, Pete and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
-  booktitle = {12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)},
-  pages     = {265--283},
-  year      = {2016},
-  publisher = {USENIX Association},
-  url       = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi}
+  title        = {TensorFlow: A System for Large-Scale Machine Learning},
+  author       = {Abadi, Mart{\'i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and Kudlur, Manjunath and Levenberg, Josh and Monga, Rajat and Moore, Sherry and Murray, Derek G. and Steiner, Benoit and Tucker, Paul and Vasudevan, Vijay and Warden, Pete and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
+  booktitle    = {12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)},
+  pages        = {265--283},
+  year         = {2016},
+  publisher    = {USENIX Association},
+  url          = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi}
 }
 
 @misc{chollet2015keras,
-  author  = {Chollet, Fran{\c{c}}ois and others},
-  title   = {Keras},
-  year    = {2015},
-  version = {3.12.1},
-  url     = {https://github.com/keras-team/keras},
-  note    = {Software}
+  author       = {Chollet, Fran{\c{c}}ois and others},
+  title        = {Keras},
+  year         = {2015},
+  version      = {3.12.1},
+  url          = {https://github.com/keras-team/keras},
+  note         = {Software}
 }
 
-
 @inproceedings{akiba2019optuna,
-  title={Optuna: A next-generation hyperparameter optimization framework},
-  author={Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
-  booktitle={Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
-  pages={2623--2631},
-  year={2019},
-  organization={ACM},
-  doi={10.1145/3292500.3330701}
+  title        = {Optuna: A next-generation hyperparameter optimization framework},
+  author       = {Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
+  booktitle    = {Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
+  pages        = {2623--2631},
+  year         = {2019},
+  organization = {ACM},
+  doi          = {10.1145/3292500.3330701}
 }
 
 @article{gundersen2018reproducible,
-  title={On reproducible AI: Towards reproducible research, open science, and digital scholarship in AI publications},
[…]
