+++ paper/paper.bib
@@ -159,19 +159,3 @@
   url          = {https://github.com/pylint-dev/pylint},
   note         = {Software}
 }
-
-@misc{pylint,
-  author       = {{pylint-dev}},
-  title        = {Pylint},
-  year         = {2026},
-  url          = {https://github.com/pylint-dev/pylint},
-  note         = {Software}
-}
-
-@misc{ruff,
-  author       = {{astral-sh}},
-  title        = {Ruff},
-  year         = {2026},
-  url          = {https://github.com/astral-sh/ruff},
-  note         = {Software}
-}
