### 2026-04-22 316859f
Fix minor BibTeX formatting for CI
