+++ paper/paper.bib
@@ -5,7 +5,7 @@
   pages        = {265--283},
   year         = {2016},
   publisher    = {USENIX Association},
-  url          = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi}
+  url          = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi},
   title        = {TensorFlow: A System for Large-Scale Machine Learning},
   author       = {Abadi, Mart{\'i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and Kudlur, Manjunath and Levenberg, Josh and Monga, Rajat and Moore, Sherry and Murray, Derek G. and Steiner, Benoit and Tucker, Paul and Vasudevan, Vijay and Warden, Pete and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
   booktitle    = {12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)},
