### 2026-04-22 d00970d
Fix minor BibTeX formatting for CI
