+++ paper/paper.bib
@@ -6,13 +6,6 @@
   year         = {2016},
   publisher    = {USENIX Association},
   url          = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi}
-  title        = {TensorFlow: A System for Large-Scale Machine Learning},
-  author       = {Abadi, Mart{\'i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and Kudlur, Manjunath and Levenberg, Josh and Monga, Rajat and Moore, Sherry and Murray, Derek G. and Steiner, Benoit and Tucker, Paul and Vasudevan, Vijay and Warden, Pete and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
-  booktitle    = {12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)},
-  pages        = {265--283},
-  year         = {2016},
-  publisher    = {USENIX Association},
-  url          = {https://www.usenix.org/conference/osdi16/technical-sessions/presentation/abadi}
 }
 
 @misc{chollet2015keras,
@@ -22,22 +15,9 @@
   version      = {3.12.1},
   url          = {https://github.com/keras-team/keras},
   note         = {Software}
-  author       = {Chollet, Fran{\c{c}}ois and others},
-  title        = {Keras},
-  year         = {2015},
-  version      = {3.12.1},
-  url          = {https://github.com/keras-team/keras},
-  note         = {Software}
 }
 
 @inproceedings{akiba2019optuna,
-  title        = {Optuna: A next-generation hyperparameter optimization framework},
-  author       = {Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
-  booktitle    = {Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
-  pages        = {2623--2631},
-  year         = {2019},
-  organization = {ACM},
-  doi          = {10.1145/3292500.3330701}
   title        = {Optuna: A next-generation hyperparameter optimization framework},
   author       = {Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
   booktitle    = {Proceedings of the 25th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
@@ -57,15 +37,6 @@
   year         = {2018},
   publisher    = {Association for the Advancement of Artificial Intelligence},
   doi          = {10.1609/aimag.v39i3.2816}
-  title        = {On reproducible AI: Towards reproducible research, open science, and digital scholarship in AI publications},
-  author       = {Gundersen, Odd Erik and Gil, Yolanda and Aha, David W.},
-  journal      = {AI Magazine},
-  volume       = {39},
-  number       = {3},
-  pages        = {56--68},
-  year         = {2018},
-  publisher    = {Association for the Advancement of Artificial Intelligence},
-  doi          = {10.1609/aimag.v39i3.2816}
 }
 
 @article{pineau2021improving,
@@ -74,23 +45,9 @@
   journal      = {arXiv preprint arXiv:2003.12206},
   year         = {2021},
   doi          = {10.48550/arXiv.2003.12206}
-  title        = {Improving reproducibility in machine learning research (a report from the NeurIPS 2019 reproducibility program)},
[…]
