### 2026-04-23 #10130 @schappag
Hi @nkrusch,

Thank you for the detailed feedback! We have addressed all the actionable comments:

- **Section headings**: Updated all section headings to use top-level `#` formatting.
- **Bibliography title**: Added `# References` at the end of `paper.md`.
- **Figure 1**: Replaced the pixelated image with a higher-resolution version in SVG format.
- **Quality control**: Added citations for the Pylint and Ruff repositories.
- **TensorFlow citation**: Updated the reference to cite the OSDI 16 proceedings version instead of the preprint.
- **Keras citation**: Added the software version number to the Keras reference.

All changes have been pushed to the repository. Please let us know if anything else needs to be addressed!
