### 2026-04-24 #10130 @nkrusch
@SiulRek, I still have a few requests:

1. remove the duplicated title

2. Figure 1: although the image quality improved, the small image size makes it difficult to read the flowchart text. Try to make the image bigger so that the flowchart text is legible.

3. Bibliography/consistency: mention versions of pylint and ruff (just like with Keras).
