### 2026-04-24 cf20cf9
Remove title from paper markdown
Removed the title from the paper's markdown header.
