### 2026-04-24 62d5dcb
Fix image size in paper.md
