### 2026-04-24 6d2a066
Update bibliography entries in paper.bib
