+++ paper/paper.bib
@@ -91,6 +91,7 @@
   author       = {{astral-sh}},
   title        = {Ruff},
   year         = {2026},
+  version      = {0.15.11},
   url          = {https://github.com/astral-sh/ruff},
   note         = {Software}
 }
@@ -99,6 +100,7 @@
   author       = {{pylint-dev}},
   title        = {Pylint},
   year         = {2026},
+  version      = {4.0.5},
   url          = {https://github.com/pylint-dev/pylint},
   note         = {Software}
 }
