### 2026-04-24 25d1b57
Update bibliography entries in paper.bib
