+++ paper/paper.bib
@@ -91,7 +91,7 @@
   author       = {{astral-sh}},
   title        = {Ruff},
   year         = {2026},
-  version      = {0.15.11},
+  version      = {0.15.5},
   url          = {https://github.com/astral-sh/ruff},
   note         = {Software}
 }
