### 2026-04-24 8547ba7
Fix image syntax in paper.md
