### 2026-04-24 8092baf
Fix image formatting in paper.md
