### 2026-04-24 #10130 @nkrusch
@SiulRek here is how you can debug the paper rendering locally:
https://joss.readthedocs.io/en/latest/paper.html#docker
