### 2026-04-24 95b84cc
Replace SVG with PDF
