### 2026-04-24 #10130 @schappag
Hi @nkrusch,
I want to apologize for triggering the editorial bot three times in a row — each time I was convinced I had finally fixed the issue with the image size. Thank you for your patience with me!

The good news: we have finally solved it. Here is what we addressed:

1. Removed the duplicated title
2. Replaced the SVG figure with a PDF. The image is now bigger.
3. Added version numbers for pylint and ruff in the bibliography, consistent with how Keras is referenced

Thanks again for your patience!
