### 2026-04-24 #10130 @nkrusch
@SiulRek Great - paper updates are now satisfactory. Next, I request that you complete the following steps.

1. Make a (new) tagged release of the software.
1. Archive the reviewed software in Zenodo, Figshare, or similar archival service.
1. Check the archival deposit has the correct metadata. This includes:
   - title (matches the paper)
   - author list (matches the paper)
   - authors' ORCIDs (match the paper)
   - license (matches the repository)
   - release version tag must match the archive version name
      For example, git tag `v1.0` would have archive name `v1.0` (displays as "Version v1.0" on Zenodo).

1. List here, as a comment:
   - the version tag
   - the archive DOI

I can then move forward with recommending accept for the submission.
