### 2026-04-24 #10130 @SiulRek
@nkrush Thank you for the guidance.

We merged the reviewed updates into the main branch, created a new tagged release, and archived the reviewed software in Zenodo.

Version tag: v1.0.0
Archive DOI: 10.5281/zenodo.19738821
